+++ crates/forge-runner/src/package_tests/with_config_resolved.rs
@@ -26,5 +26,5 @@ pub struct TestCaseResolvedConfig {
     pub expected_result: ExpectedTestResult,
     pub fork_config: Option<ResolvedForkConfig>,
     pub fuzzer_config: Option<RawFuzzerConfig>,
-    pub disable_strk_predeployment: bool,
+    pub disable_predeployed_contracts: bool,
 }