+++ crates/snforge-scarb-plugin/tests/integration/single_attributes.rs
@@ -1,5 +1,5 @@
 mod available_gas;
-mod disable_strk_predeployment;
+mod disable_predeployed_contracts;
 mod fork;
 mod fuzzer;
 mod ignore;